Add files via upload
Implement a city database using unordered lists. Each database record contains the name of the city (a string of arbitrary length) and the coordinates of the city expressed as integer x and y coordinates. Your database should allow records to be inserted, deleted by name or coordinate, and searched by name or coordinate. Another operation that should be supported is to print all records within a given distance of a specified point.

Implement the database using an array-based list implementation, and then a linked list implementation.
Collect running time statistics for each operation in both implementations.
What are your conclusions about the relative advantages and disadvantages of the two implementations?

Would storing records on the list in alphabetical order by city name speed any of the operations?

Would keeping the list in alphabetical order slow any of the operations?